destroy sfmresult class
