+++ gtsfm/common/gtsfm_data.py
@@ -1,12 +1,13 @@
 """Class to hold the tracks and cameras of a 3D scene.
+This can be the output of either data association or of bundle adjustment.
 
-Authors: Ayush Baid
+Authors: Ayush Baid, John Lambert, Xiaolong Wu
 """
 import itertools
 from typing import Dict, List, Optional, Tuple
 
 import numpy as np
-from gtsam import PinholeCameraCal3Bundler, SfmTrack
+from gtsam import PinholeCameraCal3Bundler, Pose3, SfmTrack
 
 import gtsfm.utils.graph as graph_utils
 import gtsfm.utils.logger as logger_utils
@@ -100,6 +101,20 @@ class GtsfmData:
             The camera if it is a valid one, None otherwise.
         """
         return self._cameras.get(index)
+
+    def get_camera_poses(self) -> List[Optional[Pose3]]:
+        """Getter for camera poses wTi.
+
+        This function returns the pose for all cameras (equal to number_images in GtsfmData), even if they were not
+        computed by the pipeline.
+
+        Returns:
+            camera poses as a list, each representing wTi
+        """
+        cameras = [self.get_camera(i) for i in range(self.number_images())]
+        poses = [camera.pose() if camera is not None else None for camera in cameras]
+
+        return poses
 
     def get_track(self, index: int) -> SfmTrack:
         """Getter for the track.
@@ -246,4 +261,46 @@ class GtsfmData:
             scene_reproj_errors.extend(track_errors)
 
         scene_avg_repoj_error = np.mean(scene_reproj_errors)
+
+        # DEBUG
+        scene_reproj_errors = np.array(scene_reproj_errors)
+        scene_reproj_errors.sort()
+        print('Errors after BA: ',  list(np.round(scene_reproj_errors).astype(np.uint32)))
         return scene_avg_repoj_error
+
+    def __validate_track(self, track: SfmTrack, reproj_err_thresh: float) -> bool:
+        """Validates a track based on reprojection errors and cheirality checks.
+
+        Args:
+            track: track with 3D landmark and measurements.
+            reproj_err_thresh: reprojection err threshold for each measurement.
+
+        Returns:
+            validity of the track.
+        """
+        errors, avg_reproj_error = reproj_utils.compute_track_reprojection_errors(self._cameras, track)
+        # track is valid as all measurements have error below the threshold
+        cheirality_success = np.all(~np.isnan(errors))
+        return np.all(errors < reproj_err_thresh) and cheirality_success
+
+    def filter_landmarks(self, reproj_err_thresh: float = 5) -> "GtsfmData":
+        """Filters out landmarks with high reprojection error
+
+        Args:
+            reproj_err_thresh: reprojection err threshold for each measurement.
+        """
+        # TODO: move this function to utils or GTSAM
+        filtered_data = GtsfmData(self.number_images())
+
+        # add all the cameras
+        for i in self.get_valid_camera_indices():
+            filtered_data.add_camera(i, self.get_camera(i))
+
+        for j in range(self.number_tracks()):
+            track = self.get_track(j)
+
+            if self.__validate_track(track, reproj_err_thresh):
+                filtered_data.add_track(track)
+
+        return filtered_data
+
